chore(deps): upgrade `x86_64` to v0.15

+++ src/0x02/pkg/kernel/src/memory/allocator.rs
@@ -13,7 +13,7 @@ pub fn init() {
     static mut HEAP: [u8; HEAP_SIZE] = [0; HEAP_SIZE];
 
     let heap_start = VirtAddr::from_ptr(unsafe { HEAP.as_ptr() });
-    let heap_end = heap_start + HEAP_SIZE;
+    let heap_end = heap_start + HEAP_SIZE as u64;
 
     unsafe {
         ALLOCATOR.lock().init(HEAP.as_mut_ptr(), HEAP_SIZE);
